Add authentication middleware to routes

+++ src/controllers/idea.controller.ts
@@ -1,6 +1,8 @@
 import {Request,Response,NextFunction} from 'express';
 import * as ideaServices from '../services/idea.services';
 import { ideaSchema } from '../schemas/idea.schema';
+import { getUserIdea } from '../services/idea.services';
+import { ParsedQs } from 'qs';
 
 export const createIdea = async (req: Request, res: Response, next: NextFunction) => {
     try {
@@ -26,6 +28,11 @@ export const getIdea = async (req: Request, res: Response, next: NextFunction) =
 
 export const addResources = async (req: Request, res: Response, next: NextFunction) => {
     try {
+        const user = req.body.user;
+        const userIdea= getUserIdea(req.params.id);
+        if(userIdea !==user._id){
+            throw new Error('Unauthorized');
+        }
         const idea = await ideaServices.addResources(req.params.id, req.body.resources);
         res.status(200).send(idea);
     } catch (err: any) {
@@ -35,6 +42,12 @@ export const addResources = async (req: Request, res: Response, next: NextFuncti
 
 export const updateIdea = async (req: Request, res: Response, next: NextFunction) => {
     try {
+        const user = req.body.user;
+        const userIdea= getUserIdea(req.params.id);
+        if(userIdea !==user._id){
+            throw new Error('Unauthorized');
+        }
+
         const validatedData = ideaSchema.parse(req.body);
         if (!validatedData) {
             throw new Error('Invalid data');
@@ -46,8 +59,14 @@ export const updateIdea = async (req: Request, res: Response, next: NextFunction
     }
 }
 
+
 export const deleteIdea = async (req: Request, res: Response, next: NextFunction) => {
     try {
+        const user = req.body.user;
+        const userIdea= getUserIdea(req.params.id);
+        if(userIdea !==user._id){
+            throw new Error('Unauthorized');
+        }
         const idea = await ideaServices.deleteIdea(req.params.id);
         res.status(200).send(idea);
     } catch (err: any) {
@@ -57,35 +76,9 @@ export const deleteIdea = async (req: Request, res: Response, next: NextFunction
 
 export const getIdeas = async (req: Request, res: Response, next: NextFunction) => {
     try {
-        const ideas = await ideaServices.getAllIdeas();
-        res.send(ideas);
-    } catch (err: any) {
-        next(err);
-    }
-}
-
-export const getIdeasByCategory = async (req: Request, res: Response, next: NextFunction) => {
-    try {
-        const ideas = await ideaServices.getIdeasByCategory(req.params.category);
-        res.send(ideas);
-    } catch (err: any) {
-        next(err);
-    }
-}
-
-export const getIdeasByStatus = async (req: Request, res: Response, next: NextFunction) => {
-    try {
-        const ideas = await ideaServices.getIdeasByStatus(req.params.status);
-        res.send(ideas);
-    } catch (err: any) {
-        next(err);
-    }
-}
-
-export const getIdeasByCreator = async (req: Request, res: Response, next: NextFunction) => {
-    try {
-        const ideas = await ideaServices.getIdeasByCreator(req.params.creatorId);
-        res.send(ideas);
+        let queyString = req.query;
+        const ideas = await ideaServices.getIdeas(queyString);
+        res.status(200).json(ideas);
     } catch (err: any) {
         next(err);
     }
